% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowCounts.R
\name{rowCounts}
\alias{rowCounts}
\alias{colCounts}
\alias{count}
\title{Counts the number of occurrences of a specific value or vector of values}
\usage{
rowCounts(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE,
  dim. = dim(x), ...)

colCounts(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE,
  dim. = dim(x), ...)

count(x, idxs = NULL, value = TRUE, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An NxK \code{\link[base]{matrix}} or an N * K
\code{\link[base]{vector}}.}

\item{value}{A value or vector of values to search for.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Not used.}

\item{idxs, rows, cols}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}
}
\value{
\code{rowCounts()} (\code{colCounts()}) returns an
\code{\link[base]{integer}} \code{\link[base]{vector}} of length N (K).
\code{count()} returns a scalar of type \code{\link[base]{integer}} if
the count is less than 2^31-1 (= \code{.Machine$integer.max}) otherwise
a scalar of type \code{\link[base]{double}}.  
If value is a \code{\link[base]{vector}} of J values, a N x J (K x J) matrix 
is returned with column names corresponding to the values.  
Specifying value = \code{\link[base]{NULL}} will use all unique matrix values 
(similar to \code{rowTabulates()} (\code{colTabulates()})).
}
\description{
The row- and column-wise functions take either a matrix or a vector as
input. If a vector, then argument \code{dim.} must be specified and fulfill
\code{prod(dim.) == length(x)}.  The result will be identical to the results
obtained when passing \code{matrix(x, nrow = dim.[1L], ncol = dim.[2L])},
but avoids having to temporarily create/allocate a matrix, if only such is
needed only for these calculations.
}
\examples{
x <- matrix(0:11, nrow = 4, ncol = 3)
x[2:3, 2:3] <- 2:5
x[3, 3] <- NA_integer_
print(x)

print(rowCounts(x, value = 2))
## [1]  0  1 NA  0
print(rowCounts(x, value = 2:3))
##       2  3
## [1,]  0  0
## [2,]  1  0
## [3,] NA NA
## [4,]  0  1
print(colCounts(x, value = 2))
## [1]  1  1 NA
print(colCounts(x, value = NA_integer_))
## [1] 0 0 1
print(colCounts(x, value = NULL))
##       0  1  2  3  4  7  8 11
## [1,]  1  1  1  1  0  0  0  0
## [2,]  0  0  1  1  1  1  0  0
## [3,] NA NA NA NA NA NA NA NA

print(rowCounts(x, value = 2, na.rm = TRUE))
## [1] 0 1 1 0
print(colCounts(x, value = 2, na.rm = TRUE))
## [1] 1 1 0

print(rowAnys(x, value = 2))
## [1] FALSE  TRUE  TRUE FALSE
print(rowAnys(x, value = NA_integer_))
## [1] FALSE FALSE  TRUE FALSE

print(colAnys(x, value = 2))
## [1] TRUE TRUE   NA
print(colAnys(x, value = 2, na.rm = TRUE))
## [1]  TRUE  TRUE FALSE

print(colAlls(x, value = 2))
## [1] FALSE FALSE FALSE
}
\seealso{
rowAlls, rowTabulates
}
\author{
Henrik Bengtsson. Brian Montgomery added support for a vector of
values and value = NULL.
}
\keyword{array}
\keyword{iteration}
\keyword{logic}
\keyword{univar}
