% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowTabulates.R
\name{rowTabulates}
\alias{rowTabulates}
\alias{colTabulates}
\title{Tabulates the values in a matrix by row (column).}
\usage{
rowTabulates(x, rows = NULL, cols = NULL, values = NULL, ...,
  useNames = TRUE)

colTabulates(x, rows = NULL, cols = NULL, values = NULL, ...,
  useNames = TRUE)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}}, a \code{\link[base]{logical}}, or
a \code{\link[base]{raw}} NxK \code{\link[base]{matrix}}.
Matrix `x` may only be of type \code{\link[base]{double}} if `values`
is specified and non-`NULL`, or if the values in `x` are all perfect
integers or 1/2 values. The latter allows for tabulating ranks that
may arrise from `ties.method = "average"`.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows to
operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of columns to
operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{values}{An \code{\link[base]{vector}} of J values of count. If
\code{\link[base]{NULL}}, all (unique) values are counted.}

\item{...}{Not used.}

\item{useNames}{If \code{\link[base:logical]{TRUE}} (default), names
attributes of the result are set, otherwise not.}
}
\value{
Returns a NxJ (KxJ) \code{\link[base]{matrix}} where N (K) is the
number of row (column) \code{\link[base]{vector}}s tabulated and J is the
number of values counted.
}
\description{
Tabulates the values in a matrix by row (column).
}
\details{
An alternative to these functions, is to use \code{table(x, row(x))}
and \code{table(x, col(x))}, with the exception that the latter do not
support the \code{\link[base]{raw}} data type.
When there are no missing values in \code{x}, we have that
\code{all(rowTabulates(x) == t(table(x, row(x))))} and
\code{all(colTabulates(x) == t(table(x, col(x))))}.
When there are missing values, we have that
\code{all(rowTabulates(x) == t(table(x, row(x), useNA = "always")[, seq_len(nrow(x))]))} and
\code{all(colTabulates(x) == t(table(x, col(x), useNA = "always")[, seq_len(ncol(x))]))}.
}
\examples{
x <- matrix(1:5, nrow = 10, ncol = 5)
print(x)
print(rowTabulates(x))
print(colTabulates(x))
# Count only certain values
print(rowTabulates(x, values = 1:3))


y <- as.raw(x)
dim(y) <- dim(x)
print(y)
print(rowTabulates(y))
print(colTabulates(y))
}
\author{
Henrik Bengtsson
}
\keyword{utilities}
