%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  signTabulate.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{signTabulate}
\alias{signTabulate}

\alias{signTabulate}

\title{Calculates the number of negative, zero, positive and missing values}

\usage{
signTabulate(x, ...)
}

\description{
  Calculates the number of negative, zero, positive and missing values in a \code{\link[base]{numeric}} vector.  For \code{\link[base]{double}} vectors, the number of
  negative and positive infinite values are also counted.
}

\arguments{
  \item{x}{a \code{\link[base]{numeric}} \code{\link[base]{vector}}.}
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{name}}d \code{\link[base]{numeric}} \code{\link[base]{vector}}.
}

\seealso{
  \code{\link[base]{sign}}().
}

\author{Henrik Bengtsson}


\keyword{univar}
