% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateIndices.R
\name{validateIndices}
\alias{validateIndices}
\title{Validate indices}
\usage{
validateIndices(idxs = NULL, maxIdx, allowOutOfBound = TRUE)
}
\arguments{
\item{idxs}{A \code{\link[base]{vector}} indicating subset of elements to
operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{maxIdx}{The possible max index.}

\item{allowOutOfBound}{Allow positive out of bound to indicate
\code{\link[base]{NA}}.}
}
\value{
Returns a validated integers list indicating the indices.
If some of the indices cannot be represented as an integer, the
indices are returned as doubles.
}
\description{
Computes validated positive indices from given indices.
}
\details{
\emph{WARNING: This function is defunct and will be removed in a future
version.}
}
\keyword{internal}
