% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocMatrix.R
\name{allocMatrix}
\alias{allocMatrix}
\alias{allocVector}
\alias{allocArray}
\title{Allocates an empty vector, matrix or array}
\usage{
allocMatrix(nrow, ncol, value = 0, ...)

allocVector(length, value = 0, ...)

allocArray(dim, value = 0, ...)
}
\arguments{
\item{value}{A \code{\link[base]{numeric}} scalar that all elements will
have as value.}

\item{length, nrow, ncol, dim}{\code{\link[base]{numeric}}s specifying the
dimension of the created \code{\link[base]{vector}},
\code{\link[base]{matrix}} or \code{\link[base]{array}}.}
}
\value{
Returns a \code{\link[base]{vector}}, \code{\link[base]{matrix}} and
\code{\link[base]{array}} respectively of the same data type as
\code{value}.
}
\description{
Allocates an empty vector, matrix or array faster than the corresponding
function in R.
}
\seealso{
See also \code{\link[base]{vector}}, \code{\link[base]{matrix}} and
\code{\link[base]{array}}.
}
\author{
Henrik Bengtsson
}
\keyword{internal}
\keyword{programming}
