%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{matrixStats-package}
\alias{matrixStats-package}
\alias{matrixStats}
\docType{package}

\title{Package matrixStats}


\description{
  High-performing functions operating on rows and columns of matrices, e.g. col / rowMedians(), col / rowRanks(), and col / rowSds().  Functions optimized per data type and for subsetted calculations such that both memory usage and processing time is minimized.  There are also optimized vector-based methods, e.g. binMeans(), madDiff() and weightedMedians().
}

\section{Installation}{
  To install this package, please do:
  \preformatted{
    install.packages("matrixStats")
  }
}

\section{Vignettes}{
  For an overview of the package, see the '\href{../doc/index.html}{vignettes}';
  \enumerate{
    \item Summary of functions.
  }
}

\section{How to cite this package}{
  Henrik Bengtsson (2016). matrixStats: Functions that Apply to Rows and Columns of Matrices (and to
Vectors). R package version 0.50.2. https://github.com/HenrikBengtsson/matrixStats
}

\author{Henrik Bengtsson, Hector Corrada Bravo, Robert Gentleman, Ola Hossjer, Harris Jaffee, Dongcan Jiang, Peter Langfelder}
\keyword{package}
