% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signTabulate.R
\name{signTabulate}
\alias{signTabulate}
\title{Calculates the number of negative, zero, positive and missing values}
\usage{
signTabulate(x, idxs = NULL, ...)
}
\arguments{
\item{x}{a \code{\link[base]{numeric}} \code{\link[base]{vector}}.}

\item{idxs}{A \code{\link[base]{vector}} indicating subset of elements to
operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{...}{Not used.}
}
\value{
Returns a \code{\link[base]{name}}d \code{\link[base]{numeric}}
\code{\link[base]{vector}}.
}
\description{
Calculates the number of negative, zero, positive and missing values in a
\code{\link[base]{numeric}} vector.  For \code{\link[base]{double}} vectors,
the number of negative and positive infinite values are also counted.
}
\seealso{
\code{\link[base]{sign}}().
}
\author{
Henrik Bengtsson
}
\keyword{internal}
